✨ added `series` module
♻️ add comment
