+++ tinyshift/series/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2024-2025 Lucas Leão
+# tinyshift - A small toolbox for mlops
+# Licensed under the MIT License
+
+
+from .outlier import *
+from .forecastability import *
+from .stats import *
