+++ tinyshift/stats/__init__.py
@@ -6,4 +6,3 @@
 from .utils import chebyshev_guaranteed_percentage
 from .statistical_interval import StatisticalInterval
 from .bootstrap_bca import BootstrapBCA
-from .series import hurst_exponent, hampel_filter, foreca
