--- tinyshift/stats/series.py
@@ -1,227 +1,0 @@
-# Copyright (c) 2024-2025 Lucas Leão
-# tinyshift - A small toolbox for mlops
-# Licensed under the MIT License
-
-
-from typing import Union, Tuple, List
-import numpy as np
-import scipy
-import math
-from scipy.signal import periodogram
-
-
-def hampel_filter(
-    X: Union[np.ndarray, List[float]],
-    rolling_window: int = 3,
-    factor: float = 3.0,
-    scale: float = 1.4826,
-) -> np.ndarray:
-    """
-    Identify outliers using a vectorized implementation of the Hampel filter.
-
-    The Hampel filter is a robust outlier detection method that uses the median and
-    median absolute deviation (MAD) of a rolling window to identify points that
-    deviate significantly from the local trend. This version uses vectorized operations
-    for improved performance.
-
-    Parameters
-    ----------
-    X : ndarray of shape (n_samples,) or list of float
-        Input 1D data to be filtered.
-    rolling_window : int, default=3
-        Size of the rolling window (must be odd and >= 3).
-    factor : float, default=3.0
-        Recommended values for common distributions (95% confidence):
-        - Normal distribution: 3.0 (default)
-        - Laplace distribution: 2.3
-        - Cauchy distribution: 3.4
-        - Exponential distribution: 3.6
-        - Uniform distribution: 3.9
-        Number of scaled MADs from the median to consider as outlier.
-    scale : float, default=1.4826
-        Scaling factor for MAD to make it consistent with standard deviation.
-        Recommended values for different distributions:
-        - Normal distribution: 1.4826 (default)
-        - Uniform distribution: 1.16
-        - Laplace distribution: 2.04
-        - Exponential distribution: 2.08
-        - Cauchy distribution: 1.0 (MAD is already consistent)
-        - These values make the MAD scale estimator consistent with the standard
-        deviation for the respective distribution.
-
-    Returns
-    -------
-    outliers : ndarray of shape (n_samples,)
-        Boolean array indicating outliers (True) and inliers (False).
-
-    Raises
-    ------
-    ValueError
-        If rolling_window is even or too small.
-        If input data is not 1-dimensional.
-
-    Notes
-    -----
-    The scale factor is chosen such that for large samples from the specified
-    distribution, the median absolute deviation (MAD) multiplied by the scale
-    factor approaches the standard deviation of the distribution.
-    This implementation uses vectorized operations for better performance
-    compared to the iterative version.
-    """
-
-    if rolling_window < 3:
-        raise ValueError("rolling_window must be >= 3")
-    if rolling_window % 2 == 0:
-        raise ValueError("rolling_window must be odd")
-
-    X = np.asarray(X, dtype=np.float64)
-    if X.ndim != 1:
-        raise ValueError("Input data must be 1-dimensional")
-
-    is_outlier = np.zeros(X.shape[0], dtype=bool)
-    half_window = rolling_window // 2
-    center_indices = range(half_window, X.shape[0] - half_window)
-
-    window_indices = [
-        np.arange(i - half_window, i + half_window + 1) for i in center_indices
-    ]
-    windows = X[window_indices]
-
-    medians = np.median(windows, axis=1)
-    mads = np.median(np.abs(windows - medians[:, None]), axis=1)
-    thresholds = factor * mads * scale
-
-    for i, idx in enumerate(center_indices):
-        if abs(X[idx] - medians[i]) > thresholds[i]:
-            is_outlier[idx] = True
-
-    return is_outlier
-
-
-def hurst_exponent(X: Union[np.ndarray, List[float]]) -> Tuple[float, float]:
-    """
-    Calculate the Hurst exponent using a rescaled range (R/S) analysis approach with p-value for random walk hypothesis.
-
-    The Hurst exponent is a measure of long-term memory of time series. It relates
-    to the autocorrelations of the time series and the rate at which these decrease
-    as the lag between pairs of values increases.
-
-    Parameters
-    ----------
-    X : Union[np.ndarray, List[float]]
-        Input 1D time series data for which to calculate the Hurst exponent.
-        Must contain at least 30 samples.
-
-    Returns
-    -------
-    Tuple[float, float]
-        (Hurst exponent, p-value for H=0.5 hypothesis)
-        The estimated Hurst exponent value. Interpretation:
-        - 0 < H < 0.5: Mean-reverting (anti-persistent) series
-        - H = 0.5: Geometric Brownian motion (random walk)
-        - 0.5 < H < 1: Trending (persistent) series with long-term memory
-        - H = 1: Perfectly trending series
-        p-value interpretation:
-        - p < threshold: Reject random walk hypothesis (significant persistence/mean-reversion)
-        - p >= threshold: Cannot reject random walk hypothesis
-
-    Raises
-    ------
-    ValueError
-        If input data has less than 30 samples (insufficient for reliable estimation).
-    TypeError
-        If input is not a list or numpy array.
-    """
-    X = np.asarray(X, dtype=np.float64)
-    rolling = np.diff(X)
-    size = len(rolling)
-
-    if 30 > len(X):
-        raise ValueError("Insufficient data points (minimum 30 required)")
-
-    def _calculate_rescaled_ranges(
-        rolling: np.ndarray, window_sizes: List[int]
-    ) -> np.ndarray:
-        """Helper function to calculate rescaled ranges (R/S) for each window size."""
-        r_s = np.zeros(len(window_sizes), dtype=np.float64)
-
-        for i, window_size in enumerate(window_sizes):
-            n_windows = len(rolling) // window_size
-            truncated_size = n_windows * window_size
-
-            windows = rolling[:truncated_size].reshape(n_windows, window_size)
-
-            means = np.mean(windows, axis=1, keepdims=True)
-            std_devs = np.std(windows, axis=1, ddof=1)
-            demeaned = windows - means
-            cumulative_sums = np.cumsum(demeaned, axis=1)
-            ranges = np.max(cumulative_sums, axis=1) - np.min(cumulative_sums, axis=1)
-
-            r_s[i] = np.mean(ranges / std_devs)
-
-        return r_s
-
-    def _hypothesis_test_random_walk(hurst: float, se: float, n: int) -> float:
-        """Helper function to test if Hurst exponent is significantly different from random_walk (0.5)"""
-        random_walk = 0.5
-        t_stat = (hurst - random_walk) / se
-        ddof = n - 2
-        return 2 * scipy.stats.t.sf(abs(t_stat), ddof)
-
-    max_power = int(np.floor(math.log2(size)))
-    window_sizes = [2**power for power in range(1, max_power + 1)]
-
-    rescaled_ranges = _calculate_rescaled_ranges(rolling, window_sizes)
-
-    log_sizes = np.log(window_sizes)
-    log_r_s = np.log(rescaled_ranges)
-    slope, _, _, _, se = scipy.stats.linregress(log_sizes, log_r_s)
-
-    p_value = _hypothesis_test_random_walk(slope, se, len(window_sizes))
-
-    return float(slope), float(p_value)
-
-
-def foreca(X: Union[np.ndarray, List[float]]) -> float:
-    """
-    Calculate the Forecastable Component Analysis (ForeCA) omega index for a given signal.
-
-    The omega index (ω) measures how forecastable a time series is, ranging from 0
-    (completely noisy/unforecastable) to 1 (perfectly forecastable). It is based on
-    the spectral entropy of the signal's power spectral density (PSD).
-
-    Parameters
-    ----------
-    X : Union[np.ndarray, List[float]]
-        Input 1D time series data for which to calculate the forecastability measure.
-        The signal should be stationary for meaningful results.
-
-    Returns
-    -------
-    float
-        The omega forecastability index (ω), where:
-        - ω ≈ 0: Signal is noise-like and not forecastable
-        - ω ≈ 1: Signal has strong periodic components and is highly forecastable
-
-    Notes
-    -----
-    The calculation involves:
-    1. Computing the power spectral density (PSD) via periodogram
-    2. Normalizing the PSD to sum to 1 (creating a probability distribution)
-    3. Calculating the spectral entropy of this distribution
-    4. Normalizing against maximum possible entropy
-    5. Subtracting from 1 to get forecastability measure
-
-    References
-    ----------
-    [1] Goerg (2013), "Forecastable Component Analysis" (JMLR)
-    [2] Hyndman et al. (2015), "Large unusual observations in time series"
-    [3] Manokhin (2025), "Mastering Modern Time Series Forecasting: The Complete Guide to
-        Statistical, Machine Learning & Deep Learning Models in Python", Ch. 2.4.12
-    """
-    _, psd = periodogram(X)
-    psd = psd / np.sum(psd)
-    entropy = -np.sum(psd * np.log2(psd + 1e-12))
-    max_entropy = np.log2(len(psd))
-    omega = 1 - (entropy / max_entropy)
-    return float(omega)
